requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pydantic>=2.5.0
hypothesis>=6.88.0
python-dotenv>=1.0.0